IB/rdmavt: Move MR datastructures into rvt
This patch moves the MR datastructures from hfi1 and qib and consolidates
them into rvt. For now the data structures are defined in
include/rdma/rdma_vt.h but once all MR functionality has been moved from
the drivers into rvt these should move to rdmavt/mr.h

There is also a qib specific element (list) added to rvt_mregion.  This
will go away once we move the full implementation of MR. This is the result
of an RCU bug that was not yet fixed in qib. A separate patch is being
worked directly upstream to fix qib. This won't pose an issue for rvt
because we will settle on using the HFI1 implementation and this field just
goes away anyway.

Change-Id: I98b6ea3c465548efb4223b7f2acf0c85c06dbcbe
Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>