IB/rdmavt: Remove driver specific members from qp type
In preparation for moving the queue pair data structure to rdmavt the
members of the driver specific queue pairs which are not common need to be
pushed off to a private driver structure. This structure will be available
in the queue pair once moved to rdmavt as a void pointer. This patch while
not adding a lot of value in and of itself is a prerequisite to move the
queue pair out of the drivers and into rdmavt.

The driver specific, private queue pair data structure should condense as
more of the send side code moves to rdmavt. This patch is a stepping
stone and hfi1_qp_priv and qib_qp_priv should be considered to be always
in a state of flux until the send code is merged.

Change-Id: I24a1a094191056d352c56b3c7782e7d3ccfc6010
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>