IB/hfi1: OPFN parameter negotiation
OPFN parameter negotiation allows a pair of connected RC QP's to exchange a
set of parameters in succession. The parameter exchange itself is done
using the IB compare and swap request with a special virtual address. The
request is triggered using a reserved IB work request opcode. OPFN
negotiation does not commence till the first ULP request. Because OPFN
operations are operations private to the driver, they do not generate
user completions or put the QP into error when they run out of retries.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>