IB/hfi1: TID RDMA RcvArray programming and TID allocation
This patch implements the functions for TID allocation and programming
TID RcvArray entries in hardware for kernel clients. TID entries are
managed via lists of TID groups similar to PSM. The API is designed so
that it is possible to have multiple setup/clear call pairs between a
single alloc/dealloc pair.

Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>