IB/hfi1: Prepare resource waits for dual leg
To speed data packet sending for TID RDMA WRITE request, a dual
leg protocol will be used: the first leg sends normal RC requests
while the second leg sends TID RDMA WRITE DATA pacekts. The dual
leg is susceptible to having the IOWAIT_PENDING_*  flags being
reset as resources are freed. The code attempts to set the bits
in strategic places but the strategy has locking holes because
post send could clear the pending flag out on one CPU while
an interrupt handler on another CPU frees resources and fails
to see the pending flag, stalling the QP.

This patch fixes the issue by creating a leg specific struct,
iowait_work in the iowait struct, to hold the work_struct and the
tx_list as well as a pointer to the parent iowait struct.

The hfi1_pkt_state now has an additional field to record
the current legs work structure and that is now passed to
all egress waiters to determine the leg that needs to wait
via a new iowait helper.  The APIs are adjusted to use the
new leg specific struct as required.

The case where the TID packet builder builds a verbs packet causes
a switch in the current leg's work struct and switches it back as
as appropriate.

Many new and modified helpers are added to support this change.

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>