IB/hfi1: Prepare resource waits for dual leg
Current implementation allows each qp to have only one send engine.
As such, each qp has only one list to queue prebuilt packets when send
engine resources are not available. To improve performance, it is
desired to support multiple send engines for each qp.

This patch creates the framework to support two send engines
(two legs) for each qp for the TID RDMA protocol, which can be easily
extended to support more send engines. It achieves the goal by creating
a leg specific struct, iowait_work in the iowait struct, to hold the
work_struct and the tx_list as well as a pointer to the parent iowait
struct.

The hfi1_pkt_state now has an additional field to record the current
legs work structure and that is now passed to all egress waiters to
determine the leg that needs to wait via a new iowait helper.  The
APIs are adjusted to use the new leg specific struct as required.

Many new and modified helpers are added to support this change.

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>