IB/hfi1: Add diagnostic debugfs interface
With previous version of the HFI driver, user level tools
which required access to the device's BAR made use of the
'resourceXX' files provided by the core kernel in sysfs.

However, recent distro kernels have started compiling and
distributing their kernels with CONFIG_STRICT_IO_DEVMEM
turned on.

With this option, the kernel prohibits access to device BARs
by more than one entity at the same time. This renders some
debugging tools unusable (or, at least, non-useful) since the
device cannot be operational at the same time.

The kernel does provide a workaround for this, namely the
"iomem=relaxed" boot option. However, this is not a viable
workaround for some customers.

To solve this issue, a new debugfs interface is being added,
which is designed to handle all tools that need device BAR
access. It can also be expanded to handle other tools, which
may require device memory access.

A debugfs interface has been chosen because:

  * debugfs provides a level of protection as files under it
    are only accessible by the 'root' user,
  * debugfs is meant for exactly such debugging interface, and
  * debugfs files allow for a full set of file operation
    callbacks. This, in turn, allows for processing of custom
    IOCTL commands and memory mapping of appropriate device or
    kernel memory.

Expandability is achieved by allowing user level applications to
set the "service type" they require. The "service type" is a method
that allows the driver to limit access to the device and or memory
according to the service the user application will be performing.
This includes limitting access to device memory thorugh the read and
write system calls and providing propper mmap support.

Setting of "service type" along with querying the mappable
area is done through a set of newly defined IOCTL in the RDMA
ioctl space and HFI1 command set.

The new diag interface replaces two other debugfs files - "lcb" and
"dc8051_memory".

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>