IB/{hfi1,qib}: Fix WC.byte_len calculation for UD_SEND_WITH_IMM
The work completion length for a receiving a UD send with immediate
is short by 4 bytes.

The UD receive logic incorrectly subtracts 4 bytes for immediate
value. These bytes are already included in header length and are
used to calculate header/payload split, so the result is these
4 bytes are subtracted twice, once when the header length subtracted
from the overall length and once again in the UD opcode specific
path.  Remove the extra subtraction when handling the opcode.

Fixes: 7724105686e7 ("IB/hfi1: add driver files")
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>