IB/hfi1: Add debugfs to control expansion ROM write protect
Some kernels now enable CONFIG_IO_STRICT_DEVMEM
which prevents multiple handles to PCI resource0. In order
to continue to support expansion ROM updates while the
driver is loaded, the driver must now provide an interface
to control the expansion ROM write protection.

This patch adds an exprom_wp debugfs interface that
allows the hfi1_eprom user tool to disable the expansion ROM
write protection by opening the file and writing a '1'.
The write protection is released when writing a '0' or
automatically re-enabled when the handle is closed.
The current implementation will only allow one handle
to be opened at a time across all hfi1 devices.

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Josh Collier <josh.d.collier@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>