IB/hfi1: Validate page aligned for a given virtual address
User applications can register memory regions for TID buffers
that are not aligned on page boundaries. Hfi1 is expected to
pin those pages in memory and cache the pages with mmu_rb.
The rb tree will fail to insert pages that are not aligned
correctly.

Validate whether a given virtual address is page aligned
before pinning.

Fixes: 7e7a436ecb6e ("staging/hfi1: Add TID entry program function body")
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Kamenee Arumugam <kamenee.arumugam@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>